+++ src/search/search_minimax.hpp
@@ -15,10 +15,10 @@ namespace search
 namespace details
 {
 
-int search_mini(const bitboard& bb, std::size_t depth, evaluation::evaluation eval, std::span<std::uint32_t> move_buf, const void* args) noexcept;
-int search_maxi(const bitboard& bb, std::size_t depth, evaluation::evaluation eval, std::span<std::uint32_t> move_buf, const void* args) noexcept;
+int search_mini(const bitboard& bb, std::uint8_t depth, evaluation::evaluation eval, std::span<std::uint32_t> move_buf, const void* args) noexcept;
+int search_maxi(const bitboard& bb, std::uint8_t depth, evaluation::evaluation eval, std::span<std::uint32_t> move_buf, const void* args) noexcept;
 
-inline int search_maxi(const bitboard& bb, std::size_t depth, evaluation::evaluation eval, std::span<std::uint32_t> move_buf, const void* args) noexcept
+inline int search_maxi(const bitboard& bb, std::uint8_t depth, evaluation::evaluation eval, std::span<std::uint32_t> move_buf, const void* args) noexcept
 {
     // It's a draw if our half-move clock goes too high.
     if (bb.ply_50m >= 100) [[unlikely]]
@@ -29,9 +29,9 @@ inline int search_maxi(const bitboard& bb, std::size_t depth, evaluation::evalua
 
     int ret { std::numeric_limits<int>::min() };
 
-    const std::size_t moves { generate_pseudo_legal_moves(bb, move_buf) };
+    const std::uint8_t moves { generate_pseudo_legal_moves(bb, move_buf) };
 
-    for (std::size_t i = 0; i < moves; i++)
+    for (std::uint8_t i = 0; i < moves; i++)
     {
         if (stop_search) [[unlikely]]
             return ret;
@@ -53,7 +53,7 @@ inline int search_maxi(const bitboard& bb, std::size_t depth, evaluation::evalua
     return ret;
 }
 
-inline int search_mini(const bitboard& bb, std::size_t depth, evaluation::evaluation eval, std::span<std::uint32_t> move_buf, const void* args) noexcept
+inline int search_mini(const bitboard& bb, std::uint8_t depth, evaluation::evaluation eval, std::span<std::uint32_t> move_buf, const void* args) noexcept
 {
     // It's a draw if our half-move clock goes too high.
     if (bb.ply_50m >= 100) [[unlikely]]
@@ -64,9 +64,9 @@ inline int search_mini(const bitboard& bb, std::size_t depth, evaluation::evalua
 
     int ret { std::numeric_limits<int>::max() };
 
-    const std::size_t moves { generate_pseudo_legal_moves(bb, move_buf) };
+    const std::uint8_t moves { generate_pseudo_legal_moves(bb, move_buf) };
 
-    for (std::size_t i = 0; i < moves; i++)
+    for (std::uint8_t i = 0; i < moves; i++)
     {
         if (stop_search) [[unlikely]]
             return ret;
@@ -90,7 +90,7 @@ inline int search_mini(const bitboard& bb, std::size_t depth, evaluation::evalua
 
 }
 
-inline int search_minimax(const bitboard& bb, std::size_t depth, std::span<std::uint32_t> move_buf, evaluation::evaluation eval, const void* args_eval)
+inline int search_minimax(const bitboard& bb, std::uint8_t depth, std::span<std::uint32_t> move_buf, evaluation::evaluation eval, const void* args_eval)
 {
     return bb.is_black_to_play() ? details::search_mini(bb, depth, eval, move_buf, args_eval) : details::search_maxi(bb, depth, eval, move_buf, args_eval);
 }
